+++ src/main/java/zajecia08/carswithaircondition/Car.java
@@ -3,7 +3,7 @@ package zajecia08.carswithaircondition;
 public class Car extends Vehicle {
     private boolean isConditionAir;
 
-    public Car(String name, float capacityTank, float mediumFuelConsumption, boolean isConditionAir) {
+    public Car(String name, double capacityTank, double mediumFuelConsumption, boolean isConditionAir) {
         super(name, capacityTank, mediumFuelConsumption);
         this.isConditionAir = isConditionAir;
     }
@@ -19,6 +19,7 @@ public class Car extends Vehicle {
 
     @Override
     String info() {
+
         return super.info() + " klimatyzacja włączona: " + isConditionAir;
     }
 }
